Added a preloader effect
The effect lasts 2 seconds in total. It's delayed for 1.5 seconds and then it fades out in 0.5 seconds.

I used a 2 second timeout in combination with it. I tried to make the effect disappear initially after the page loads, but I didn't end up figuring it out, so I used a manually set delay of 2 seconds.